rename example corpus files/class
